- [ZBX-4175] sync to trunk; manually resolve conflicts in frontends/php/api/classes/class.chostgroup.php
git-svn-id: svn://svn.zabbix.com/branches/dev@22511 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-4175/src/libs/zbxcommon/alias.c
@@ -78,7 +78,7 @@ void	alias_list_free()
 	aliasList = NULL;
 }
 
-void	alias_expand(const char *orig, char *expanded, int exp_buf_len)
+void	alias_expand(const char *orig, char *expanded, size_t exp_buf_len)
 {
 	ALIAS	*alias;
